+++ nycschools/snapshot.py
@@ -13,8 +13,16 @@
 # You should have received a copy of the License along with this program.
 # If not, see <http://www.gnu.org/licenses/>.
 # ==============================================================================
+
+from requests import Session
+import pandas as pd
+import os
+
+from concurrent.futures import ThreadPoolExecutor, as_completed
+
+
 from . import config
-from nycschools.dataloader import load
+from .dataloader import load
 
 
 def load_snapshots():
@@ -25,3 +33,37 @@ def load_snapshots():
     """
     filename = config.urls["snapshots"].filename
     return load(filename)
+
+
+def get_snapshot_schools():
+    """
+    Returns the list of schools in the snapshot data.
+    """
+
+    results = []
+
+
+    def read_year(year, file):
+        try:
+            url = f"https://tools.nycenet.edu/vue-assets/static/data/sqs/{year}/{file}.csv"
+            x = pd.read_csv(url)
+            x["year"] = year
+            return x
+        except:
+            print("no data,", year)
+            return pd.DataFrame()
+
+
+    for year in range(2016, 2023):
+        results.append(read_year(year, "school_info"))
+        results.append(read_year(year, "school_info_pk"))
+
+
+    doe = pd.concat(results, axis=0)
+    return doe
+    # path = os.path.join(config.data_dir, "snapshot_school_list.csv")
+    # doe.to_csv("snapshot_school_list.csv", index=False)
+    return doe
+    
+
+